image start

+++ src/main/java/ua/lviv/cinema/service/CreditCardService.java
@@ -32,7 +32,7 @@ public interface CreditCardService {
      * @param creditCard - credit card to be charged
      * @param amount - amount to be charded
      */
-    void chargeCreditCard(CreditCard creditCard, BigDecimal amount) throws Exception;
+    void chargeCreditCard(CreditCard creditCard, int amount) throws Exception;
